Extract geometry primitives into dedicated package
