+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "geom",
     "win32",
     "d2d",
     "game",
