+++ geom/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "geom"
+version = "0.1.0"
+edition = "2021"
+authors = ["Connor Power <connor@connorpower.com>"]
+
+[features]
+default = []
+# Enables conversions to and from Win32 foundational geometry types.
+win32 = ["windows/Win32_Foundation"]
+# Enables conversions to and from Direct2D geometry types.
+d2d = ["windows/Win32_Graphics_Direct2D_Common"]
+
+[dependencies]
+num-traits = "0.2.15"
+
+[dependencies.windows]
+version = "0.42.0"
+optional = true
+
+[dev-dependencies]
+pretty_assertions = "1.3.0"
