+++ win32/Cargo.toml
@@ -5,11 +5,12 @@ edition = "2021"
 authors = ["Connor Power <connor@connorpower.com>"]
 
 [dependencies]
+geom = { path = "../geom", features = ["win32"] }
+
 bitvec = "1.0.1"
 deku = "0.15"
 thiserror = "1.0.37"
 lazy_static = "1.4.0"
-num-traits = "0.2.15"
 maplit = "1.0.2"
 parking_lot = "0.12.1"
 paste = "1.0.9"
